backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
easyocr==1.7.1
openai==1.13.3
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.6.4
spacy==3.7.4
reportlab==4.1.0
